requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
requests==2.31.0
python-dotenv==1.0.0
psycopg[binary]==3.2.3
gunicorn==21.2.0
pytz
